Refactor onboarding page into reusable components
